working on backing up and restoring via json (rewriting it)

+++ src/rest/v1.0/lib/Database.php
@@ -296,15 +296,9 @@ class Database{
 
 		$keepGoing = "true";
 
-		// read the file if present
-		$handle = @fopen($_SERVER['BACKUPDATA'], 'r+');
+		$fileName = $obj['fileName'];
 
-		// create the file if needed
-		if($handle === null || $handle === false){
-		    $handle = fopen($_SERVER['BACKUPDATA'], 'w+');
-		}
-
-		var_dump($handle);
+		$finalStuff = array();
 
 		while($keepGoing == "true"){
 			$results = matchAll200($from);
@@ -315,29 +309,7 @@ class Database{
 						$stuffToKeep = $results['hits']['hits'];
 
 						for($x = 0; $x < count($stuffToKeep); $x++){
-							if($handle){
-							    // seek to the end
-							    fseek($handle, 0, SEEK_END);
-
-							    // are we at the end of is the file empty
-							    if(ftell($handle) > 0){
-							        // move back a byte
-							        fseek($handle, -1, SEEK_END);
-
-							        // add the trailing comma
-							        fwrite($handle, ',', 1);
-
-							        // add the new json string
-							        fwrite($handle, json_encode($stuffToKeep[$x]['_source']) . ']');
-							    }
-							    else{
-							        // write the first event inside an array
-							        fwrite($handle, json_encode(array($stuffToKeep[$x]['_source'])));
-							    }
-
-						        // close the handle on the file
-						        fclose($handle);
-							}
+							array_push($finalStuff, $stuffToKeep[$x]['_source']);	
 						}
 
 						$from += 200;
@@ -352,11 +324,15 @@ class Database{
 			}
 		}
 
-		//file_put_contents($_SERVER['BACKUPDATA'], json_encode($finalStuff));
+		file_put_contents($_SERVER['BACKUPJSONPATH'] . $fileName . " - backup", json_encode($finalStuff));
 
-		$stuff = file_get_contents($_SERVER['SERVICECREDS']);
+		$keepServiceCreds = $obj['keepServiceCreds'];
 
-		file_put_contents($_SERVER['SERVICECREDSBACKUP'], $stuff);
+		if($keepServiceCreds == "true"){
+			$stuff = file_get_contents($_SERVER['SERVICECREDS']);
+
+			file_put_contents($_SERVER['SERVICECREDSBACKUP'], $stuff);
+		}
 
 		$wipeCurrentData = $obj['wipeCurrentData'];
 
@@ -401,9 +377,13 @@ class Database{
 		$data = file_get_contents($_SERVER['BACKUPDATA']);
 		$data = json_decode($data, true);
 
-		for($x = 0; $x < count($data); $x++){
-			$result = $this->writeObject($data[$x]);
+		foreach($data as $key => $value){
+			$result = $this->writeObject($data[$key]);
 		}
+
+		/*for($x = 0; $x < count($data); $x++){
+			$result = $this->writeObject($data[$x]);
+		}*/
 
 		if(isset($obj)){
 			if(isset($obj['deleteBackupFile'])){
